# docs/04-api-and-integrations/OPENAPI-SPEC.yaml
openapi: 3.0.3
info:
  title: Plataforma de Eventos e Ingressos API
  description: API REST para comercialização de ingressos, controle de assentos sem double-booking, importação de catálogos TMDb/Ticketmaster e validação criptográfica na portaria.
  version: 1.0.0
servers:
  - url: http://localhost:3000/api
    description: Servidor de Desenvolvimento Local
paths:
...
  /seats/reserve:
    post:
      summary: Bloqueio atômico de assentos numerados (Anti-Double Booking com TTL 10m)
      tags: [Reservas]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [seatIds]
              properties:
                eventId:
                  type: string
                sectorId:
                  type: string
                seatIds:
                  type: array
                  items:
                    type: string
                  example: ["cuid_seat_1", "cuid_seat_2"]
      responses:
        '200':
          description: Assentos bloqueados com sucesso (retorna reservationId e expiresAt)
        '409':
          description: Conflito de concorrência - um ou mais assentos já foram reservados
...
  /tickets/{id}/cancel:
    post:
      summary: Cancela voluntariamente um ingresso ativo, liberando o assento/cota e estornando o valor
      tags: [Ingressos]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: ID único do ingresso a ser cancelado
      responses:
        '200':
          description: Ingresso cancelado com sucesso e assento devolvido ao estoque
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  ticketId:
                    type: string
                  status:
                    type: string
                    example: "CANCELLED"
                  seatRestored:
                    type: string
                    nullable: true
                    example: "A1"
                  refundAmount:
                    type: number
                    example: 250.0
                  cancelledAt:
                    type: string
                    format: date-time
        '400':
          description: Ingresso já utilizado, já cancelado ou evento já ocorrido
        '401':
          description: Não autenticado
        '403':
          description: Acesso negado - apenas o titular do ingresso pode cancelá-lo
        '404':
          description: Ingresso não encontrado
...